document unsupported 83+ OS versions, fix broken latex
After seeing another report of HT2 not working on TI-83+ running OS
1.13, I've decided to officially drop support for those legacy OS
versions. Upgrading to 1.19 isn't overly difficult, and has no added
downsides. If somebody wants to work on a proper fix, please do.

Also, the maintainers of the `fourier` CTAN package decided to change
the mapping of the "danger" symbol for some reason. Fixed the TeX
source accordingly.